ME build system: New version of the ME build system.
It moves so3 in the refso3 to keep the ME context at one place while
still allowing to have all the generic code in one place only.
It also uses one build.sh script in the root folder which manages the
build of the MEs.

+++ ME/deploy.sh
@@ -11,8 +11,12 @@ sudo rm -rf fs/*
 if [ "$1" != "clean" ]; then
     if [ "$1" != "" ]; then
         ME_to_deploy="../../ME/$1/target/so3virt.itb"
-        sudo cp -rf $ME_to_deploy fs/
-        echo "$1 deployed"
+        if [ -f ${ME_to_deploy} ]; then
+            sudo cp -rf $ME_to_deploy fs/
+            echo "$1 deployed"
+        else
+            echo "The ITB you want to deploy is not present."
+        fi
     else
         echo "No ME specified to be deployed!"
     fi
